+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 1.x branch
+## 1.1 branch
+### 1.1.0
+* `ThumbManager::clear()` and `ThumbManager::clearAll()` can return `null`;
+* updated for `php` 7.1 and `phpunit` 8.
+
 ## 1.0 branch
 ### 1.0.1
 * fixed lowest dependencies;